% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.party.C5.0.R
\name{as.party.C5.0}
\alias{as.party.C5.0}
\title{Convert C5.0 object to party format}
\usage{
\method{as.party}{C5.0}(obj, trial = 0, ...)
}
\arguments{
\item{obj}{A \code{C5.0} class object'}

\item{trial}{An integer for the specific tree to plot.}

\item{...}{Not currently used.}
}
\value{
A \code{party} object
}
\description{
Convert C5.0 object to party format
}
\keyword{internal}
