% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C50-package.R
\docType{package}
\name{C50-package}
\alias{C50}
\alias{C50-package}
\title{C50: C5.0 Decision Trees and Rule-Based Models}
\description{
C5.0 decision trees and rule-based models for pattern recognition that extend the work of Quinlan (1993, ISBN:1-55860-238-0).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://topepo.github.io/C5.0/}
  \item \url{https://github.com/topepo/C5.0/}
  \item Report bugs at \url{https://github.com/topepo/C5.0/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{mxkuhn@gmail.com}

Authors:
\itemize{
  \item Ross Quinlan (Author of imported C code)
}

Other contributors:
\itemize{
  \item Steve Weston [contributor]
  \item Mark Culp [contributor]
  \item Nathan Coulter [contributor]
  \item RuleQuest Research (Copyright holder of imported C code) [copyright holder]
  \item Rulequest Research Pty Ltd. (Copyright holder of imported C code) [copyright holder]
}

}
\keyword{internal}
